requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-WTF==0.12
awesome-slugify==1.6.5